requirements.txt
fastapi
uvicorn
pydantic
pyjwt 